% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Graph a country's population}
\usage{
CountryPopulation(countryName)
}
\arguments{
\item{countryName}{An input string}
}
\value{
A graph
}
\description{
This function will display the graph the years 1950 to 2020 of an inputed
country's population
}
\examples{
CountryPopulation('United States of America')
}
